Backend/website/requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
coverage==7.6.4
Django==5.1.2
django-allauth==65.1.0
django-cors-headers==4.6.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
flake8==7.1.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
isodate==0.7.2
itypes==1.2.0
Jinja2==3.1.4
kavenegar==1.1.2
lxml==5.3.0
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
simplejson==3.19.3
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
zeep==4.3.1
